% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtraj.R
\name{buildtraj}
\alias{buildtraj}
\title{Wrapper of flexmix}
\usage{
buildtraj(
  Rdat,
  formula = as.formula(cbind(A, 1 - A) ~ time),
  J,
  id = "id",
  family = "binomial",
  control = list(iter.max = 1000, minprior = 0),
  ...
)
}
\arguments{
\item{Rdat}{sample data to build trajectory groups. These data are in long format.
Each row represent a person time, column A is a binary data.}

\item{formula}{designate the formula to model the longitudinal variable of interest.}

\item{J}{an integer to choose the number of trajectory groups.}

\item{family}{designate the type of distribution "gaussian", "binomial", "poisson" and "gamma".}

\item{control}{object of class FLXcontrol.}

\item{\dots}{to add supplementary functions.}
}
\value{
\item{dpost}{Posterior probability.}

\item{model}{Fitted trajectory model.}
}
\description{
Call the package flexmix to construct trajectory groups
}
\examples{
obsdata = gendatatraj()
Rdat =longtowide(obsdata = obsdata, varying = 1:5)
head(Rdat)
res.traj = buildtraj(Rdat = Rdat, k=3,formula = cbind(A,1-A) ~ time, id=id)
head(res.traj$dpost)
}
\author{
Awa Diop, Denis Talbot
}
