% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_datatraj.R
\name{gendatatraj}
\alias{gendatatraj}
\title{Data Simulation for Trajectory Analysis}
\usage{
gendatatraj(
  n1 = 250,
  n2 = 350,
  n3 = 400,
  beta01 = -5,
  beta02 = -0.15,
  beta03 = -0.01,
  beta11 = -1,
  beta12 = 0.5,
  beta13 = 5,
  set.seed = 355
)
}
\arguments{
\item{n1}{sample size of the first subgroup.}

\item{n2}{sample size of the second subgroup.}

\item{n3}{sample size of the third subgroup.}

\item{beta01}{intercept for the first subgroup.}

\item{beta02}{intercept for the second subgroup.}

\item{beta03}{intercept for the third subgroup.}

\item{beta11}{slope for the first subgroup.}

\item{beta12}{slope for the second subgroup.}

\item{beta13}{slope for the third subgroup.}

\item{set.seed}{to add a seed.}
}
\value{
\item{Obsdata}{Wide format data.}
}
\description{
Example of longitudinal data with three hidden subgroups to perform LCGA/GBTM.
}
\examples{
Obsdata = gendatatraj()
head(Obsdata)
}
\author{
Awa Diop, Denis Talbot
}
