% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gform_countermeans.R
\name{gform_countermeans}
\alias{gform_countermeans}
\alias{sub_gform}
\title{Counterfactual means for g-formula.}
\usage{
sub_gform(formula, V, L, A, Y, s, time, obsdata)
}
\arguments{
\item{formula}{specification of the model for the outcome to be fitted.}

\item{V}{baseline covariates.}

\item{L}{time-varying covariates.}

\item{A}{time-varying treatment.}

\item{Y}{outcome variable.}

\item{s}{number of measuring times.}

\item{time}{measuring times.}

\item{obsdata}{observed data in wide format.}
}
\value{
\item{list_gform_countermeans}{Counterfactual means obtained with g-formula.}
}
\description{
Get the counterfactual means for the g-formula.
}
\examples{
Obswidedata = longtowide(Obsdata = gendatrajMSM(n=500), idvar = "ID", timevar = "Time");
formula = paste0("Y.2013~", paste0("Statins.", c(2011:2013),collapse = "+"), "+",
paste0("BMI.", c(2011:2013),collapse = "+"),"+",
paste0("Hyper.", c(2011:2013),collapse = "+"),"+",
"Age.2011 + Sex.2011", collapse = "+")
Y = "Y.2013 "
A = "Statins."
L = c("Hyper.", "BMI.")
V = c("Age.", "Sex.")
s=3
time = c(2011,2012,2013)
res.gform = sub_gform(dat=Obswidedata, formula = formula, Y=Y, A=A,L=L,V=V,s=3, time=time)
res.gform$counter_means
}
\author{
Awa Diop, Denis Talbot
}
