% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gformula.R
\name{gformula}
\alias{gformula}
\title{Counterfactual means for g-formula.}
\usage{
gformula(
  formula,
  baseline,
  covariates,
  treatment,
  outcome,
  ntimes_interval,
  obsdata
)
}
\arguments{
\item{formula}{specification of the model for the outcome to be fitted.}

\item{baseline}{name of the baseline covariates.}

\item{covariates}{list of the names of the time-varying covariates.}

\item{treatment}{name of the time-varying treatment.}

\item{ntimes_interval}{length of a time-interval (s).}

\item{obsdata}{observed data in wide format.}

\item{Y}{outcome variable.}
}
\value{
\item{list_gform_countermeans}{Counterfactual means obtained with g-formula.}
}
\description{
Get the counterfactual means for the g-formula.
}
\author{
Awa Diop, Denis Talbot
}
