% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotraj.R
\name{ggplotraj}
\alias{ggplotraj}
\title{ggplotraj}
\usage{
ggplotraj(
  Rdat = NULL,
  dpost = NULL,
  trajdat = NULL,
  formula = as.formula(A ~ time + class),
  trt = "A",
  time = "time",
  id = "id",
  class = "class",
  FUN = "mean",
  ...
)
}
\arguments{
\item{Rdat}{an object produced by the function \code{"longtowide"}.}

\item{dpost}{matrix of the posterior probabilities and the trajectory groups.}

\item{trt}{name of the time-varying treatment.}

\item{time}{name of the variable measurements of time.}

\item{id}{name of the id variable.}

\item{class}{name of the trajectory groups.}

\item{FUN}{specify what statistics to display, by default calculate the mean.}

\item{\dots}{to add supplementary functions.}

\item{Trajdat}{merged datasets containing observed data in long format and trajectory groups.}
}
\value{
\item{p}{Plot the trajectory groups using the observed treatment.}
}
\description{
Use \code{"ggplot2"} to plot trajectory groups produced by the function \code{"buildtraj"} using the observed treatment.
}
\examples{
Obsdata = gendatatraj()
Rdat = longtowide(Obsdata = Obsdata, varying = 1:5)
head(Rdat)
res.traj = buildtraj(Rdat = Rdat, k=3,formula = cbind(A,1-A) ~ time, id="id")
dpost = res.traj$dpost
head(dpost)
ggplotraj(Rdat = Rdat, dpost = dpost, formula = A ~ time + class,
trt = "A",time = "time",id="id",class = "class", FUN = "mean")
}
\author{
Awa Diop, Denis Talbot
}
