% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPW.R
\name{IPW}
\alias{IPW}
\title{Inverse of Probability Weighting}
\usage{
IPW(
  numerator = c("stabilized", "unstabilized"),
  id = id,
  V = V,
  L = L,
  A = A,
  C = FALSE,
  s = s,
  time = time,
  censor = NULL,
  obsdata = obsdata_msm
)
}
\arguments{
\item{id}{name of the column for unique identifiant.}

\item{V}{baseline covariates.}

\item{L}{time-varying covariates.}

\item{A}{time-varying treatment.}

\item{C}{logical value TRUE/FALSE to include or not a censoring variable.}

\item{s}{number of measuring times.}

\item{time}{measuring times.}

\item{censor}{name of the censoring variable.}

\item{obsdata}{observed data in wide format.}
}
\value{
\item{IPW}{Stabilized and unstabilized inverse of probabilities with and without censoring}
}
\description{
Compute stabilized and unstabilized with and without censor weights.
}
\examples{
dat = gendatTrajMSM(n=500, Censor=FALSE)
V <- c("Age","Sex")
L <- c("Hyper","BMI")
time <- "Time"
A <- "Statins"
Censor = "C"
sw = IPW(numerator = "stabilized",id= "ID", V = c("Age","Sex"),
     L = c("Hyper","BMI"),
     time ="Time",
     A=c("Statins"),obsdata = dat)
summary(sw)
}
\author{
Awa Diop, Denis Talbot
}
