% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longtowide.R
\name{longtowide}
\alias{longtowide}
\title{Reshape from long to wide format.}
\usage{
longtowide(obsdata, idvar, timevar)
}
\arguments{
\item{obsdata}{the data to reshape in a long format.}

\item{varying}{varying columns.}
}
\value{
\item{long_dat}{Long format data.}
}
\description{
To convert observed data into a wide format for the g-formula and pooled ltmle.
}
\examples{
widedata = longtowide(obsdata = gendatrajMSM(n=500), idvar = "ID", timevar = "Time")
head(widedata)
}
\author{
Awa Diop, Denis Talbot
}
