% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooledltmle_countermeans.R
\name{pooledltmle_countermeans}
\alias{pooledltmle_countermeans}
\alias{sub_pooled_ltmle}
\title{Counterfactual means for a Pooled LTMLE}
\usage{
sub_pooled_ltmle(
  formula,
  id,
  Y,
  A,
  L,
  V,
  s,
  time,
  time_name,
  trajname,
  obsdata,
  trajmodel
)
}
\arguments{
\item{formula}{specification of the model for the outcome to be fitted.}

\item{id}{name of the column for unique identifiant.}

\item{Y}{outcome variable.}

\item{A}{time-varying treatment.}

\item{L}{time-varying covariates.}

\item{V}{baseline covariates.}

\item{s}{number of measuring times.}

\item{time}{measuring times.}

\item{obsdata}{observed data in wide format.}
}
\value{
\item{list_pooledltmle_countermeans}{Counterfactual means and influence functions with the pooled ltmle.}
\item{D}{Influence functions}
}
\description{
function to estimate counterfactual means for a pooled LTMLE.
}
\examples{
obsdata_msm <- genobsdata_msm(n=2500)
V <- "V"
L <- "L"
time <- 1:3
A <- "A"
Y <- "Y"
}
\author{
Awa Diop, Denis Talbot
}
