% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicTraj.R
\name{predicTraj.}
\alias{predicTraj.}
\alias{predicTraj}
\title{Predict trajectory groups for deterministic treatment regimes}
\usage{
predicTraj(id, s, trt, time_name, trajmodel)
}
\arguments{
\item{id}{name of the column for unique identifiant.}

\item{s}{number of measuring times per interval.}

\item{trt}{name of the time-varying treatment.}

\item{time_name}{name of the measuring times.}

\item{trajmodel}{trajectory model built with the observed treatment.}

\item{name}{of the id column variable.}
}
\description{
function to predict trajectory groups for deterministic treatment regimes
used with gformula and pooled LTMLE.
}
\author{
Awa Diop, Denis Talbot
}
