% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split Observed data into multiple subsets}
\usage{
split_data(obsdata, K, s, A, L, Y)
}
\arguments{
\item{obsdata}{observed data in wide format.}

\item{K}{total length of follow-up.}

\item{s}{number of measuring times per interval.}

\item{A}{time-varying treatment.}

\item{L}{time-varying covariates.}

\item{V}{baseline covariates.}
}
\value{
\item{df}{all subsets, time-intervals.}
}
\description{
function to split the data into multiple subsets of size s each
one corresponding to one time-interval.
}
\author{
Awa Diop Denis Talbot
}
