% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajhrmsm_ipw.R
\name{trajhrmsm_ipw}
\alias{trajhrmsm_ipw}
\title{History Restricted MSM and Latent Class of Growth Analysis estimated with IPW.}
\usage{
trajhrmsm_ipw(
  degree_traj = c("linear", "quadratic", "cubic"),
  numerator = c("stabilized", "unstabilized"),
  identifier,
  baseline,
  covariates,
  treatment,
  outcome,
  var_cov,
  include_censor = FALSE,
  ntimes_interval,
  total_followup,
  time,
  time_values,
  family = "poisson",
  censor = censor,
  number_traj,
  obsdata,
  weights = NULL,
  treshold = 0.999
)
}
\arguments{
\item{degree_traj}{to specify the polynomial degree for modelling the time-varying treatment.}

\item{numerator}{To choose between stabilized and unstabilized weights.}

\item{identifier}{name of the column for unique identifiant.}

\item{baseline}{name of baseline covariates.}

\item{covariates}{names of time-varying covariates in a wide format.}

\item{treatment}{name of time-varying treatment.}

\item{outcome}{name of the outcome variable.}

\item{var_cov}{names of the time-varying covariates in a long format.}

\item{include_censor}{Logical, if TRUE, includes censoring.}

\item{ntimes_interval}{length of a time-interval (s).}

\item{total_followup}{total length of follow-up.}

\item{time}{name of the time variable.}

\item{time_values}{values of the time variable.}

\item{family}{specification of the error distribution and link function to be used in the model.}

\item{censor}{name of the censoring variable.}

\item{number_traj}{number of trajectory groups.}

\item{obsdata}{data in a long format.}

\item{weights}{a vector of estimated weights. If NULL, the weights are computed by the function.}

\item{treshold}{for weight truncation.}
}
\value{
A list containing the following components:
  \itemize{
  \item{results_hrmsm_ipw}{Estimates of LCGA-HRMSM.}
  \item{result_coef_boot}{Estimates obtained with bootstrap.}
  \item{restraj}{Fitted trajectory model.}
  \item{mean_adh}{Mean adherence per trajectory group.}
  }
}
\description{
Estimate parameters of LCGA-HRMSM Using a Pooled LTMLE.
}
\examples{
obsdata_long = gendata(n = 1000, format = "long", total_followup = 8,
timedep_outcome = TRUE,  seed = 945)
baseline_var <- c("age","sex")
years <- 2011:2018
variables <- c("hyper", "bmi")
covariates <- lapply(years, function(year) {
paste0(variables, year)})
treatment_var <- paste0("statins", 2011:2018)
var_cov <- c("statins","hyper", "bmi","y")
reshrmsm_ipw <- trajhrmsm_ipw(degree_traj = "linear", numerator = "stabilized",
identifier = "id", baseline = baseline_var,
covariates = covariates, treatment = treatment_var,
outcome = "y", var_cov= var_cov,include_censor = FALSE,
 ntimes_interval = 6,total_followup = 8, time = "time", time_values = 2011:2018,
family = "poisson", number_traj = 3, obsdata = obsdata_long, treshold = 0.999)
reshrmsm_ipw$res_trajhrmsm_ipw
}
\author{
Awa Diop, Denis Talbot
}
