% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajhrmsm_pltmle.R
\name{trajhrmsm_pltmle}
\alias{trajhrmsm_pltmle}
\title{History Restricted MSM and Latent Class of Growth Analysis estimated with a Pooled LTMLE.}
\usage{
trajhrmsm_pltmle(
  degree_traj = c("linear", "quadratic", "cubic"),
  treatment,
  covariates,
  baseline,
  outcome,
  ntimes_interval,
  total_followup,
  time,
  time_values,
  identifier,
  var_cov,
  number_traj = 3,
  family = "poisson",
  obsdata,
  treshold = 0.99
)
}
\arguments{
\item{degree_traj}{to specify the polynomial degree for modelling the time-varying treatment.}

\item{treatment}{name of time-varying treatment.}

\item{covariates}{names of time-varying covariates in a wide format.}

\item{baseline}{name of baseline covariates.}

\item{ntimes_interval}{length of a time-interval (s).}

\item{total_followup}{total length of follow-up.}

\item{time}{name of the time variable.}

\item{time_values}{measuring times.}

\item{identifier}{name of the column for unique identifiant.}

\item{var_cov}{names of the time-varying covariates in a long format.}

\item{number_traj}{number of trajectory groups.}

\item{family}{specification of the error distribution and link function to be used in the model.}

\item{obsdata}{data in a long format.}

\item{formula}{specification of the model for the outcome to be fitted for a binomial or gaussian distribution.}

\item{censor}{name of the censoring variable.}

\item{rep}{number of repetition for the bootstrap.}
}
\value{
A list containing the following components:
  \itemize{
  \item{results_hrmsm_pltmle}{Estimates of LCGA-HRMSM.}
  \item{result_coef_boot}{Estimates obtained with bootstrap.}
  \item{restraj}{Fitted trajectory model.}
  \item{mean_adh}{Mean adherence per trajectory group.}
  }
}
\description{
Estimate parameters of LCGA-HRMSM Using a Pooled LTMLE.
}
\examples{
obsdata_long = gendata(n = 1000, format = "long",
total_followup = 8, timedep_outcome = TRUE,  seed = 945)
baseline_var <- c("age","sex")
years <- 2011:2018
variables <- c("hyper", "bmi")
covariates <- lapply(years, function(year) {
  paste0(variables, year)})
treatment_var <- paste0("statins", 2011:2018)
var_cov <- c("statins","hyper", "bmi","y")
respltmle = trajhrmsm_pltmle(degree_traj = "linear", treatment = treatment_var,
covariates = covariates, baseline = baseline_var,
outcome = paste0("y", 2016:2018),var_cov = var_cov, ntimes_interval = 6,
total_followup = 8, time = "time",time_values = years, identifier = "id",
number_traj = 3, family = "poisson", obsdata = obsdata_long)
respltmle$results_hrmsm_pltmle
}
\author{
Awa Diop Denis Talbot
}
