% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajMSM_gform.R
\name{trajMSM_gform}
\alias{trajMSM_gform}
\title{Parametric g-formula}
\usage{
trajMSM_gform(
  formula = paste0("Y~", paste0("A", 1:s, collapse = "+"), "+", paste0("L", 1:s,
    collapse = "+"), "+", V, collapse = "+"),
  rep = 50,
  id,
  V,
  L,
  A,
  Y,
  s,
  time,
  time_name,
  J = 2,
  trajmodel,
  obsdata
)
}
\arguments{
\item{formula}{specification of the model for the outcome to be fitted.}

\item{rep}{number of repetitions for the bootstrap.}

\item{id}{name of the column for unique identifiant.}

\item{V}{baseline covariates.}

\item{L}{time-varying covariates.}

\item{A}{time-varying treatment.}

\item{s}{number of measuring times.}

\item{time}{measuring times.}

\item{time_name}{name of the variable time.}

\item{trajmodel}{trajectory model built with the observed treatment.}

\item{obsdata}{observed data in wide format.}
}
\value{
\item{results_msm_gform}{Estimates of a LCGA-MSM with g-formula.}
}
\description{
function to estimate parameters of a MSM-LCGA using g-formula
 and bootstrap to get standard errors.
}
\examples{
obsdata_msm <- genobsdata_msm(n=2500)
V <- "V"
L <- "L"
time <- 1:3
A <- "A"
Y <- "Y"
Adat_long <- widetolong(obsdata = obsdata_msm[, c("A1","A2","A3")], varying = 1:3)
res.traj = buildtraj(Rdat = Adat_long, J=3,formula = cbind(A,1-A) ~ time, id= "id")
obsdata_msm$class = res.traj$dpost$class
trajMSM_gform(formula = paste0("Y~", paste0("A", 1:s,collapse = "+"), "+",
                              paste0("L", 1:s,collapse = "+"),"+",
                              V, collapse = "+"),rep = 50,
             id = id, V = V,L = L,A = A,Y=Y,s=s,time = time,time_name = "time",J=3,
             family = family,trajmodel = res.traj$model, obsdata = obsdata_msm)
}
\author{
Awa Diop Denis Talbot
}
