% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajMSM_IPW.R
\name{trajMSM_IPW}
\alias{trajMSM_IPW}
\title{Marginal Structural Model and Latent Class of Growth Analysis estimated with IPW}
\usage{
trajMSM_IPW(
  formula1 = as.formula("Y ~ class"),
  formula2 = as.formula("Surv(Y,event) ~ class"),
  numerator = c("stabilized", "unstabilized"),
  id,
  V,
  L,
  A,
  Y,
  C = FALSE,
  censor = NULL,
  s,
  time,
  family = c("binomial", "gaussian", "survival"),
  J,
  obsdata,
  weights = NULL
)
}
\arguments{
\item{formula1}{specification of the model for the outcome to be fitted for a binomial or gaussian distribution.}

\item{formula2}{specification of the model for the outcome to be fitted for a survival outcome.}

\item{id}{name of the column for unique identifiant.}

\item{V}{baseline covariates.}

\item{L}{time-varying covariates.}

\item{A}{time-varying treatment.}

\item{censor}{name of the censoring variable.}

\item{s}{number of measuring times.}

\item{time}{measuring times.}

\item{family}{specification of the error distribution and link function to be used in the model.}

\item{J}{number of trajectory groups.}

\item{weights}{a vector of estimated weights. If NULL, the weights are computed by the function.}
}
\value{
\item{IPW}{Stabilized and unstabilized inverse of probabilities}
}
\description{
Combine Marginal Structural Model and Latent Class of Growth Analysis
}
\examples{
obsdata_msm <- genobsdata_msm(n=2500)
V <- "V"
L <- "L"
time <- 1:3
A <- "A"
Y <- "Y"
Adat_long <- widetolong(obsdata = obsdata_msm[, c("A1","A2","A3")], varying = 1:3)
res.traj = buildtraj(Rdat = Adat_long, J=3,
formula = cbind(A,1-A) ~ time, id= "id")
obsdata_msm$class = res.traj$dpost$class
trajMSM_IPW(formula1 = as.formula("Y ~ class"), numerator = "stabilized",
          id = id, V = V,L = L,A = A,Y=Y,
          C = FALSE,s = s, time = time, family = "binomial",
          obsdata = obsdata_msm)
}
\author{
Awa Diop, Denis Talbot
}
