% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajMSM_pooled_ltmle.R
\name{trajMSM_pooled_ltmle}
\alias{trajMSM_pooled_ltmle}
\title{Pooled LTMLE}
\usage{
trajMSM_pooled_ltmle(
  formula = paste0("Y~", paste0("A", 1:s, collapse = "+"), "+", paste0("L", 1:s,
    collapse = "+"), "+", paste0("V", 1:s, collapse = "+"), collapse = "+"),
  id,
  V,
  L,
  A,
  Y,
  s,
  time,
  time_name,
  J = 2,
  trajmodel,
  obsdata
)
}
\arguments{
\item{formula}{specification of the model for the outcome to be fitted.}

\item{id}{name of the column for unique identifiant.}

\item{L}{covariates.}

\item{A}{time-varying treatment.}

\item{s}{number of measuring times per interval.}

\item{time}{measuring times.}

\item{J}{an integer to choose the number of trajectory groups.}

\item{trajmodel}{trajectory model built with the observed treatment.}

\item{obsdata}{observed data in wide format.}
}
\value{
\item{results_msm_pooledltmle}{Estimates of a LCGA-MSM with pooled LTMLE.}
}
\description{
function to estimate parameters of a MSM-LCGA using pooled LTMLE
 with influence functions or boostrap to estimate standard errors.
}
\examples{
obsdata_msm <- genobsdata_msm(n=2500)
V <- "V"
L <- "L"
time <- 1:3
A <- "A"
Y <- "Y"
Adat_long <- widetolong(obsdata = obsdata_msm[, c("A1","A2","A3")], varying = 1:3)
res.traj = buildtraj(Rdat = Adat_long, J=3,formula = cbind(A,1-A) ~ time, id= "id")
obsdata_msm$class = res.traj$dpost$class
trajMSM_pooled_ltmle(formula = paste0("Y~", paste0("A", 1:s,collapse = "+"), "+",
                              paste0("L", 1:s,collapse = "+"),"+",
                              V, collapse = "+"),
             id = id, V = V,L = L,A = A,Y=Y,s = s,time = time,time_name = "time",J=3,
             family = family,trajmodel = res.traj$model, obsdata = obsdata_msm)
}
\author{
Awa Diop, Denis Talbot
}
