% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widetolong.R
\name{widetolong}
\alias{widetolong}
\title{Reshape from wide to long format.}
\usage{
widetolong(obsdata, varying)
}
\arguments{
\item{obsdata}{the data to reshape in a long format.}

\item{varying}{varying columns.}
}
\value{
\item{widetolong_dat}{Long format data.}
}
\description{
To convert data generated by the function gen_datatraj into a long format.
}
\examples{
longdata = widetolong(obsdata = gendatatraj(), varying = 1:5)
head(longdata)
}
\author{
Awa Diop, Denis Talbot
}
