% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajhrmsm_gform.R
\name{trajhrmsm_gform}
\alias{trajhrmsm_gform}
\title{g-formula without SuperLearner}
\usage{
trajhrmsm_gform(
  degree_traj = c("linear", "quadratic", "cubic"),
  rep = 50,
  treatment,
  covariates,
  baseline,
  outcome,
  ntimes_interval,
  total_followup,
  time,
  time_values,
  identifier,
  var_cov,
  number_traj = 3,
  family = "poisson",
  obsdata
)
}
\arguments{
\item{degree_traj}{to specify the polynomial degree for modelling the time-varying treatment.}

\item{rep}{number of repetition for the bootstrap.}

\item{treatment}{name of time-varying treatment.}

\item{covariates}{names of time-varying covariates in a wide format.}

\item{baseline}{name of baseline covariates.}

\item{ntimes_interval}{length of a time-interval (s).}

\item{total_followup}{total length of follow-up.}

\item{time}{name of the time variable.}

\item{time_values}{measuring times.}

\item{identifier}{name of the column for unique identifiant.}

\item{var_cov}{names of the time-varying covariates in a long format.}

\item{number_traj}{number of trajectory groups.}

\item{family}{specification of the error distribution and link function to be used in the model.}

\item{obsdata}{data in a long format.}

\item{formula}{specification of the model for the outcome to be fitted for a binomial or gaussian distribution.}

\item{censor}{name of the censoring variable.}
}
\value{
\item{trajMSM_gform }{g-formula}
}
\description{
function to estimate parameters of a HRMSM-LCGA using g-formula
 and bootstrap to get standard errors.
}
\examples{
obsdata_long = gendata(n = 1000, format = "long", total_followup = 8,
timedep_outcome = TRUE,  seed = 945)
baseline_var <- c("age","sex")
years <- 2011:2018
variables <- c("hyper", "bmi")
covariates <- lapply(years, function(year) {
paste0(variables, year)})
treatment_var <- paste0("statins", 2011:2018)
var_cov <- c("statins","hyper", "bmi")
reshrmsm_gform = trajhrmsm_gform(degree_traj = "linear", rep=2 ,
treatment = treatment_var,covariates = covariates, baseline = baseline_var,
outcome = "y",var_cov = var_cov, ntimes_interval = 6, total_followup = 8,
 time = "time",time_values = years, identifier = "id",
number_traj = 3, family = "poisson", obsdata = obsdata_long)
reshrmsm_gform$results_hrmsm_gform
}
\author{
Awa Diop Denis Talbot
}
